Move courierListUtils into utils folder
